Add log-changed-deps command to plugin
log-changed-deps produced a summary of the changes in project
dependencies between the rebar.config.lock file in the working
directory and the revision specified by rev=REV (defaults to HEAD).

This is a useful summary to review new dependency versions as well as
something that could be used as the commit log for a new locked dep
release.

The command logs new, removed, and changed deps. For changed deps, a
summary of the commits on that dep between the last locked version and
the newly locked version is provided.

The output can be captured in a file using log=FILENAME