Add commit-release and tag-release commands
commit-release (blindly) adds the current rebar.config.lock and
rel/reltool.config files and creates a git commit with log message
based on the output of the log-changed-deps command.

tag-release is a very simple wrapper to `git tag -a -m` setting the
tag name to the release version in reltool.config and setting a
standardized tag message.