Respect order of rebar.config files when generating lock file
Before this patch, lock-deps would generate an alphabetized listing
of dependencies in the lock file. This was a side-effect of using
filelib:wildcard to list the entries in the dep dir.

With this patch, the relative order of dependencies as they appear in
the various rebar.config files is (mostly) preserved when generating
the lock file via lock-deps. When a dep is shared, only the first
occurance is kept. As a result, deps that are shared will be listed
out of order from the point of view of some rebar.config files.

Preserving the ordering in the lock file is important when making use
of parse transforms. For example, if a project uses lager's parse
transform it must be built before projects making use of the
transform.

Since generated lock files change as a result of this patch, the major
version of the plugin has been increased. This is a conservative
approach. It should be the case that the reordering is back
compatible.

Thanks to Oliver Ferrigni for identifying this issue, providing an
initial fix prototype, and reviewing this work.