WIP: deal with missing deps by ignoring them
Some projects use rebar.config.script to remove items from the deps
list (and others could use it to add items). Presently, the lock deps
plugin naively reads rebar.config without respecting
rebar.config.script.

In the case of gproc, the "right" thing is to ignore the missing deps
which is what this work-in-progress patch does.

Should this be the behavior? Should we bite off proper handling of
rebar.config.script?